Refactor SavedPlans.vue to be its own component. Also show the button only when the user is signed in
